requirements.txt
fastapi>=0.110.0,<0.130.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.3,<3.2
aiosqlite>=0.20.0
python-dotenv>=1.0.1
python-multipart>=0.0.20
feedparser>=6.0.0
httpx>=0.27.0
boto3>=1.35.0
anthropic>=0.40.0
markdown>=3.5
slowapi>=0.1.9
pillow>=10.0
